Add core-platform-legacy to supported api surfaces
This is needed because `external/icu4j` creates two versions of the
`core-platform` API, one containing stable APIs and one containing
legacy APIs. It is not clear exactly what the difference is but as
they pass different `--show*-annotation` options to Metalava they need
to have their own `<api-surface>` element in the configuration.

Bug: 391554590
Test: m checkapi
Change-Id: Ie02abd17a289261fe47bfb9f2da0ca472f065d0d