Convert checkapi and update-api to Soong
By converting the phony generation logic from Make to Soong.

This change also leads to some differences in the dependencies of
`checkapi` and `droidcore` phony targets. `checkapi` and `droidcore`
currently depend on `<droidstubs name>-api-lint` and `checkapi` phony
targets respectively, but this change modifies the behavior by making
`checkapi` and `droidcore` depend on the api lint timestamp and the
checkapi timestamps directly.

Test: diff aninja query of checkapi
Bug: 389719475
Change-Id: I46aa2de03e54b121ac7627c0a752026e65dc6933