Install apex symbol file
This is a partial revert of aosp/3308088. We need the symbol file
to still be installed, because a later ctx.InstallSymlink file adds
a dependency on the it.

We removed that install initially so that it didn't get installed in
filesystems, to replace that, just exclude files installed to the
"apex" partition from filesystems.

Bug: 393617060
Test: cd bionic; mm
Change-Id: Id4abe966c630d1e356575386af233eab92849934