Store name of boot image profile providing module in config
The logic in getProfileProviderForApex depends on having a list of
all apexes that depend on the art-bootclasspath-fragment module,
which will no longer be available once the apex info top down
mutator is removed.  Since only the art-bootclasspath-fragment
module needs to generate a boot image profile, put the name of
the module in the boot image config.

Bug: 372543712
Test: all soong tests pass
Change-Id: I920e3f1d85e8d74ff8e047310e10504bcdde06db