Introduce all_aconfig_declarations_extension module type
The newly introduced module type is used as an extension of
all_aconfig_declarations singleton in order to provide the ability of
non-platform api surfaces to depend on the aconfig flags
in api finalization. Although all_aconfig_declarations is a singleton
module, the newly introduced extension module is not a singleton; this
is because there can be multiple non-platform api surfaces that requires
access to all aconfig flags for api finalization.

Test: m nothing --no-soong-only
Bug: 391344176
Change-Id: Ib26b8e2abf9cec0d69ee647b5cc56142c42375c2