droidstubs: pass in --current-version <major.minor> to metalava
Instead of just passing the API level (or just the major SDK version) to
metalava via --current-version, pass in both major and minor version
(e.g. --current-version 36.1).

Ideally ApiLevel should be updated to support minor versions (and
potentially renamed SdkVersion), but that work falls outside the scope
of this CL which is intended to unblock work to support an SDK with a
non-zero minor version.

Bug: 392838750
Bug: 397644333
Test: m nothing
Ignore-AOSP-First: minor SDK development takes place on internal main
Change-Id: I17151b357dd06176dbaab3f9bdea29affe5c5011