Remove special case that installs apex variants of test modules
b/146995717 installed apex variants of test modules, claiming that the
tests won't be included in an apex, but the only use of this appears
to be the tests in the com.android.art.testing apex.  Those tests are
packaged in an apex, and the special case to also install them results
in duplicate install rules that then require the workaround in
b/395617594 to hide them from the platform.  Remove both the special
case and the workaround.

This causes the platform variants of the tests in the
com.android.art.testing apex that are installe to /data/nativetest
to come from the platform variant instead of the apex variant, and
causes libart-broken.so to no longer be installed in /data/nativetest.

Bug: 395617594
Test: compare installs-${TARGET_PRODUCT}.mk
Change-Id: Ie2fc70508ababd0bb6965bdc66450d71f2a342f7