cc: Add cc_rustlibs_for_make to bundle rust_ffi_static
Adds a new module cc_rustlibs_for_make to whole-link the generated
staticlib for a cc_library_static that would otherwise wait for a final
link to be generated. This is intended to support bundling
rust_ffi_static libraries for use in Make as they're no longer true
staticlibs and instead rlibs.

Bug: 393511828
Bug: 396666286
Test: m <Make static executable with rust_ffi_static deps>
Change-Id: I158ecf9b7493ecb259f3732af0db821a828efd5b