Build system_server.zip in soong
For soong-only builds.

The make code was wrapped in 3 conditions: that dexpreopt was enabled,
that we're not only preopting the art boot image, and that
PRODUCT_USES_DEFAULT_ART_CONFIG is true. This code only has the first 2
of those conditions, but that's because if
PRODUCT_USES_DEFAULT_ART_CONFIG is false, dexpreopt is disabled:
https://cs.android.com/android/platform/superproject/main/+/main:build/make/core/dex_preopt_config.mk;l=9;drc=6a8fff595ba4cbb87b7e124fd630f548c5c4e2cb

Fixes: 398036389
Test: Diff'd system_server.zip before/after this change
Change-Id: Ie316c56bf07296372ed89e7ee2f0f38360fe24d7