Fix incrementality issues for root_filesystem_config.txt
This file is generated by running fs_config on
`out/soong/.intermediates/.../ROOT` directory assembled for
target_files.zip, but this was not declared as implicit input. This was
causing incorrect results for clean builds. To fix this, a duplicate
rule is created to assemeble the ROOT/ directory in a separate
intermediates. An output timestamp
file is created for the aforementioned directory, and this timestamp
file will be an implicit input for the fs_config call.

Test: rm -rf out/soong/.intermediates/build/soong/fsgen && # build
target_files.zip

Change-Id: Id3f942c285d95148eba78c6010b42ca061b69072