Remove ab_update=true from dynamic_partition_list.txt
When building with Make, this entry is present only in misc_info.txt and
not not in dynamic_partition_list.txt. This entry gets added to the
first file via `dump-super-image-info`.

This CL ports this behavior to Soong by moving this entry out of
`dumpDynamicPartitionInfo` and into `buildMiscInfo`

Bug: 399788100
Test: Verified that this entry is present in Soong's misc_info.txt
Test: Verified that this entry is absent in Soong's
dynamic_partition_info.txt

Change-Id: I98384510c7f47c403ca5d7edbf863e513c9d503d